Actually fix wire.h for linter
God I detest the style rules sometimes.

+++ megaavr/libraries/Wire/src/Wire.h
@@ -67,21 +67,22 @@ class TwoWire: public Stream {
     void setClock(uint32_t);
 
     void begin();
+    // all attempts to make these look prettier were rejected by astyle, and it's not worth disabling linting over.
     void begin(uint8_t  address, bool receive_broadcast, uint8_t second_address);
-    void begin(int16_t  address, bool receive_broadcast, uint8_t second_address) {
-      begin(  (uint8_t) address,      receive_broadcast,         second_address);
+    void begin(int      address, bool receive_broadcast, uint8_t second_address) {
+      begin((uint8_t) address, receive_broadcast, second_address);
     }
     void begin(uint8_t  address, bool receive_broadcast) {
-      begin(            address,      receive_broadcast, 0);
+      begin(address, receive_broadcast, 0);
     }
-    void begin(int16_t  address, bool receive_broadcast) {
-      begin(  (uint8_t) address,      receive_broadcast, 0);
+    void begin(int      address, bool receive_broadcast) {
+      begin((uint8_t) address, receive_broadcast, 0);
     }
     void begin(uint8_t  address) {
-      begin(            address, 0, 0);
+      begin(address, 0, 0);
     }
-    void begin(int16_t  address) {
-      begin(  (uint8_t) address, 0, 0);
+    void begin(int      address) {
+      begin((uint8_t) address, 0, 0);
     }
 
     void end();
@@ -89,8 +90,8 @@ class TwoWire: public Stream {
     void endSlave(void);
 
     void beginTransmission(uint8_t address);
-    void beginTransmission(int16_t address) {
-      beginTransmission(  (uint8_t)address);
+    void beginTransmission(int     address) {
+      beginTransmission((uint8_t)address);
     }
     uint8_t endTransmission(bool);
     uint8_t endTransmission(void) {
@@ -123,10 +124,10 @@ class TwoWire: public Stream {
     inline size_t write(long          n) {
       return      write((uint8_t)     n);
     }
-    inline size_t write(uint16_t      n) {
+    inline size_t write(uint          n) {
       return      write((uint8_t)     n);
     }
-    inline size_t write(int16_t       n) {
+    inline size_t write(int           n) {
       return      write((uint8_t)     n);
     }
     using Print::write;
